# Conflicts:
#	src/main/java/com/neuedu/hismedicalsystem/model/mapper/PatientMapper.java
#	src/main/resources/application.properties
#	src/main/resources/mapper/RegistrationMapper.xml
#	src/sql/his.sql

+++ src/main/java/com/neuedu/hismedicalsystem/model/mapper/PatientMapper.java
@@ -21,5 +21,5 @@ public interface PatientMapper {
 
     void insertHomepage(int regid);
 
-    FrontPage getFrontPageByRegid(long regid);
+    FrontPage getFrontPageByRegid(int regid);
 }
